Context API changes
I'm currently working on a project that is a ripoff of the well-known Netflix video hosting service. It also provides a dashboard for administrators where they can check statistics for the website. It is a React app with a Node JS and Express backend, MongoDB as the database, and JWT as the authentication method.